+++ tests/explicit_access_modifier/jsfmt.spec.ts
@@ -1,1 +1,4 @@
-runSpec(__dirname, ["apex"], { apexExplicitAccessModifier: true });
+runSpec(__dirname, ["apex"], {
+  apexExplicitAccessModifier: true,
+  apexSortModifiers: true,
+});